Fog update

+++ Assets/Scripts/Core/EntityProperties.cs
@@ -7,7 +7,7 @@ public abstract class EntityProperties : ScriptableObject
 {
     public string Name;
     public float Hp;
-    public float visionDistance;
+    public int visionDistance;
 
     public float Damage;
     public DmgType DamageType;
